data_collector/requirements.txt
Flask==3.0.0
Flask-APScheduler==1.12.4
mysql-connector-python==8.2.0
grpcio==1.60.0
grpcio-tools==1.60.0
requests==2.31.0
protobuf==4.21.12
python-dotenv==1.0.0